Fixed bad update_interval type

+++ opendps/opendps.h
@@ -32,7 +32,7 @@
 /** Max number of parameters to a function */
 #define OPENDPS_MAX_PARAMETERS  (8)
 
-extern uint16_t opendps_update_interval;
+extern uint32_t opendps_update_interval;
 
 /**
  * @brief      Enable specified function
